mouse trail
trail following cursor position (looks better imo)
