+++ Assets/Prefabs/Mouse/mouse_trail.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f1f3b0f4b74f9eb458316d3f15492e8e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
