+++ Assets/Scenes/Game Design/Screen Navigation/jr/initial_game_load.unity
@@ -200,63 +200,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 142134862}
   m_CullTransparentMesh: 1
---- !u!1001 &688573176
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_RootOrder
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128896, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9078208608291128909, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
-      propertyPath: m_Name
-      value: Cursor_with_trail
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 7f41349a02a0ae04fbd38327c289b69d, type: 3}
 --- !u!1 &863039130
 GameObject:
   m_ObjectHideFlags: 0
@@ -950,3 +893,60 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1960260269}
   m_CullTransparentMesh: 1
+--- !u!1001 &9055791934674499688
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936720, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 22769140527936733, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
+      propertyPath: m_Name
+      value: Cursor_with_trail
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: f1f3b0f4b74f9eb458316d3f15492e8e, type: 3}
